refactor: generalize _CafeImageProgressIndicator to ImageProgressIndicator
